Fix user defined executors options
Hey TJ, thanks so much for your Advent of Neovim series on YouTube! It's been the best thing to watch this month!

Anyways, I just noticed this part that you missed in part 8 of the series, and I think without this fix it would try to use a user-defined Lua executor for JavaScript and Python!

Keep up the great work, this has been really enlightening and inspiring me to try make my own plugin!